+++ piglegsurgeryweb/uploader/admin.py
@@ -2,10 +2,11 @@ from django.contrib import admin
 
 # Register your models here.
 
-from .models import UploadedFile, BitmapImage
+from .models import UploadedFile, BitmapImage, Owner
 
 admin.site.register(UploadedFile)
 admin.site.register(BitmapImage)
+admin.site.register(Owner)